move user query tests
