+++ api-js/src/queries/user/__tests__/find-user-by-username.test.js
@@ -4,14 +4,14 @@ const { ArangoTools, dbNameFromFile } = require('arango-tools')
 const { graphql, GraphQLSchema, GraphQLError } = require('graphql')
 const { toGlobalId } = require('graphql-relay')
 const { setupI18n } = require('@lingui/core')
-const { makeMigrations } = require('../../migrations')
-const { userRequired } = require('../auth')
-const { createQuerySchema } = require('../queries')
-const { cleanseInput } = require('../validators')
-const { createMutationSchema } = require('../mutations')
-const { userLoaderByKey, userLoaderByUserName } = require('../loaders')
-const englishMessages = require('../locale/en/messages')
-const frenchMessages = require('../locale/fr/messages')
+const { makeMigrations } = require('../../../../migrations')
+const { userRequired } = require('../../../auth')
+const { createQuerySchema } = require('../..')
+const { cleanseInput } = require('../../../validators')
+const { createMutationSchema } = require('../../../mutations')
+const { userLoaderByKey, userLoaderByUserName } = require('../../../loaders')
+const englishMessages = require('../../../locale/en/messages')
+const frenchMessages = require('../../../locale/fr/messages')
 
 describe('given the findUserByUsername query', () => {
   let query,
